+++ tests/LinqContraband.Tests/Analyzers/LC009_MissingAsNoTracking/MissingAsNoTrackingMessageTests.cs
@@ -3,8 +3,7 @@ using Microsoft.CodeAnalysis.Testing;
 using AnalyzerTest = Microsoft.CodeAnalysis.CSharp.Testing.CSharpAnalyzerTest<
     LinqContraband.Analyzers.LC009_MissingAsNoTracking.MissingAsNoTrackingAnalyzer,
     Microsoft.CodeAnalysis.Testing.Verifiers.XUnitVerifier>;
-using VerifyCS = Microsoft.CodeAnalysis.CSharp.Testing.XUnit.AnalyzerVerifier<
-    LinqContraband.Analyzers.LC009_MissingAsNoTracking.MissingAsNoTrackingAnalyzer>;
+using VerifyCS = LinqContraband.Tests.Analyzers.LC009_MissingAsNoTracking.ReturnedEntitiesOptInVerifier;
 
 namespace LinqContraband.Tests.Analyzers.LC009_MissingAsNoTracking;